Add compatibility with AVR-GCC installed via Boards Manager
Previously the path to the copy of AVR-GCC included with the Arduino IDE
installation was assumed. However Boards Manager installation of AVR-GCC
causes a different path structure that is not compatible with this
assumption. This can occur in multiple ways:
- Boards Manager installation of Marlin_AT90USB causes installation of
AVR-GCC
- Manual installation of Marlin_AT90USB will be broken by Boards Manager
installation of a hardware package that has AVR-GCC as one of its
dependencies.